+++ packages/org/public/forms/org_unit-fm.js
@@ -1,6 +1,0 @@
-exports.formCode = {
-    initUBComponent: function () {
-  	},
-    isEditable: function() {return false;}
-
-};