+++ packages/cdn/CHANGELOG.md
@@ -15,6 +15,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.4.43] - 2020-11-10
+### Changed
+- `cdn_profession.meta`: added `defaultView: false` for attribute `description`
+
+### Removed
+-`cdn_profession-fm.def`: better to use autoform
+
 ## [5.4.42] - 2020-11-08
 ## [5.4.41] - 2020-11-08
 ## [5.4.40] - 2020-11-05