+++ packages/org/CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.4.13] - 2020-11-10
+### Changed
+- `org_unit-fm`, `org_staffunit-fm`, `org_profession-fm`, `org_department-fm`: rewritten to vue
+
 ## [5.4.12] - 2020-11-08
 ## [5.4.11] - 2020-11-08
 ## [5.4.10] - 2020-11-05